pyproject.toml
[tool.poetry]
name = "til_bigquery"
version = "0.1.0"
description = "BigQuery integration with Pydantic models"
authors = ["Hynek Blaha <hynek.blaha@timeisltd.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
pydantic = "^1.7.3"
google-cloud-bigquery = "^2.7.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
mypy = "^0.800"
pylint = "^2.6.0"
pre-commit = "^2.10.0"
Faker = "^5.8.0"

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
known_first_party = []
known_third_party = []

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
